refs #36739 - Security improvement Last fixes to protext XSS and SQL Injection Readme Update - contributions & docker

+++ controllers/front/webhook.php
@@ -60,7 +60,7 @@ class YounitedpayWebhookModuleFrontController extends ModuleFrontController
             $this->endResponse('Accès refusé : ' . $response->getReasonPhrase());
         }
 
-        $idCart = Tools::getValue('id_cart');
+        $idCart = (int) Tools::getValue('id_cart');
 
         if ($idCart === false) {
             $this->endResponse('Error, no Cart Id Provided');
